Unify derivation model and make serialization fallible
Independent attribute parsing allowed serialization, deserialization, and
sizing to choose different defaults for the same field. This normalizes it to
use one shared model such that we can ensure all three paths use the same
defaults. Additionally, to be more truthful, we allow writing to fail using
errors instead of panicking or resulting in bad data. For example, when the size
 of collections was larger than their length field, we would've just written
incorrect data (client would read too little due to overflow). While the user
_should_ ensure they only create valid packets, we currently have no way to
prevent it. As such, the first place we can actually do something is when the
element is serialized. Given we _know_ this can happen, and there are a few
cases we know of, this is now explicitly modeled as errors.

+++ crates/skrillax-serde/src/error.rs
@@ -10,7 +10,7 @@ pub enum SerializationError {
     #[error("I/O error when serialize/deserializing packet at field {0}. {1:?}")]
     FieldIoError(&'static str, io::Error),
     #[error("The enum {1} does not have a variation for value {0}")]
-    UnknownVariation(usize, &'static str),
+    UnknownVariation(u64, &'static str),
     #[error("Could not convert bytes to a string")]
     StringParsingFailed(#[from] FromUtf8Error),
     #[error("Could not convert bytes to a utf16 string")]
